Switching From a Website
Scrapping the website idea as it will take too long to do and is not very realistic in terms of my current capabilities. Swift will be much easier and will allow me to use some of my pre-existing skills with firebase and ones gained from previous projects. Also just a side note added sir to the repository.